requirements.txt
numpy~=2.3.4
tensorflow~=2.20.0
kagglehub~=0.3.13
keras~=3.10.0
fastapi~=0.95.2
ptl~=0.3.0
requests~=2.31.0
python-dotenv~=1.0.1
uvicorn[standard]~=0.23.2